requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
passlib[bcrypt]
python-jose
dnspython
jinja2
python-multipart